+++ packages/carp_communication_package/lib/communication.dart
@@ -11,7 +11,7 @@ import 'dart:convert';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:another_telephony/telephony.dart';
 import 'package:call_e_log/call_log.dart';
-import 'package:device_calendar/device_calendar.dart' as cal;
+import 'package:device_calendar_plus/device_calendar_plus.dart' as cal;
 import 'package:crypto/crypto.dart';
 import 'package:permission_handler/permission_handler.dart';